requirements.txt
# UI
streamlit>=1.43.2

# LangChain Core
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.19
langchain-huggingface>=0.3.1

# Google Gemini
google-genai>=1.5.0
langchain-google-genai>=2.1.0

# Embeddings
sentence-transformers>=5.2.0

# Document Processing
pymupdf>=1.27.1
python-docx>=1.1.2

# Web Research
duckduckgo-search>=7.3.0
wikipedia>=1.4.0
# Utilities
python-dotenv>=1.0.1
requests>=2.32.0
# Vector Database
endee>=0.1.0